safari-600.7-branch/LayoutTests/fast/mediastream/RTCPeerConnection-onnegotiationneeded-expected.txt

Tests RTCPeerConnection onnegotiationneeded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Got a stream.
PASS onNegotiationNeeded was called.
PASS stream.getAudioTracks().length is 1
PASS stream.getAudioTracks().length is 0
PASS onNegotiationNeeded was called when track was removed.
PASS stream.getAudioTracks().length is 0
PASS stream.getAudioTracks().length is 1
PASS onNegotiationNeeded was called when track was added.
PASS successfullyParsed is true

TEST COMPLETE
